requirements-gpu.txt
requests==2.18.4
numpy==1.14.5
keras==2.2.0
tensorflow-gpu==1.4.0
pillow==5.2.0
matplotlib==2.2.2
opencv-python==3.4.1.15
configparser==3.5.0
graphviz==0.8.4
pydot==1.2.4
xmltodict==0.11.0
mxnet-gpu==1.2.1